requirements.txt
uvicorn
fastapi
tortoise-orm
